mvd stuff to uint64

+++ product.go
@@ -55,7 +55,7 @@ type Product struct {
 // The options provided by Shopify
 type ProductOption struct {
 	ID        uint64      `json:"id,omitempty"`
-	ProductID int      `json:"product_id,omitempty"`
+	ProductID uint64      `json:"product_id,omitempty"`
 	Name      string   `json:"name,omitempty"`
 	Position  int      `json:"position,omitempty"`
 	Values    []string `json:"values,omitempty"`
