+++ variant.go
@@ -29,7 +29,7 @@ type VariantServiceOp struct {
 
 // Variant represents a Shopify variant
 type Variant struct {
-	ID                   int              `json:"id,omitempty"`
+	ID                   uint64              `json:"id,omitempty"`
 	ProductID            int              `json:"product_id,omitempty"`
 	Title                string           `json:"title,omitempty"`
 	Sku                  string           `json:"sku,omitempty"`
@@ -66,7 +66,7 @@ type VariantsResource struct {
 }
 
 // List variants
-func (s *VariantServiceOp) List(productID int, options interface{}) ([]Variant, error) {
+func (s *VariantServiceOp) List(productID uint64, options interface{}) ([]Variant, error) {
 	path := fmt.Sprintf("%s/%d/variants.json", productsBasePath, productID)
 	resource := new(VariantsResource)
 	err := s.client.Get(path, resource, options)
@@ -74,13 +74,13 @@ func (s *VariantServiceOp) List(productID int, options interface{}) ([]Variant, 
 }
 
 // Count variants
-func (s *VariantServiceOp) Count(productID int, options interface{}) (int, error) {
+func (s *VariantServiceOp) Count(productID uint64, options interface{}) (int, error) {
 	path := fmt.Sprintf("%s/%d/variants/count.json", productsBasePath, productID)
 	return s.client.Count(path, options)
 }
 
 // Get individual variant
-func (s *VariantServiceOp) Get(variantID int, options interface{}) (*Variant, error) {
+func (s *VariantServiceOp) Get(variantID uint64, options interface{}) (*Variant, error) {
 	path := fmt.Sprintf("%s/%d.json", variantsBasePath, variantID)
 	resource := new(VariantResource)
 	err := s.client.Get(path, resource, options)
@@ -88,7 +88,7 @@ func (s *VariantServiceOp) Get(variantID int, options interface{}) (*Variant, er
 }
 
 // Create a new variant
-func (s *VariantServiceOp) Create(productID int, variant Variant) (*Variant, error) {
+func (s *VariantServiceOp) Create(productID uint64, variant Variant) (*Variant, error) {
 	path := fmt.Sprintf("%s/%d/variants.json", productsBasePath, productID)
 	wrappedData := VariantResource{Variant: &variant}
 	resource := new(VariantResource)
@@ -106,6 +106,6 @@ func (s *VariantServiceOp) Update(variant Variant) (*Variant, error) {
 }
 
 // Delete an existing product
-func (s *VariantServiceOp) Delete(productID int, variantID int) error {
+func (s *VariantServiceOp) Delete(productID uint64, variantID int) error {
 	return s.client.Delete(fmt.Sprintf("%s/%d/variants/%d.json", productsBasePath, productID, variantID))
 }
